Oxygen Saturation Percentage #196 Differing KG Weights for Same LBS Weight #116 Update Treatment Complete Button Name #199 Correct Amputation Designations #63 Vital Signs Entry Prior to Applied Template #21 Patient Designations #135
For both the Patient Information Table and the Vital Signs History, if there is no BSA to calculate (e.g. no BSA_Method/Weight_Formula) then the BSA columns will display "N/A" with a hover over of "Not Available" rather than the previous display of "0.00"
Additionally in the Patient Information Table, if a template is not currently applied and active for a patient, then the "Clinical Trial" display will be blank (currently it displays "Not a clinical trial" even if a template has never been applied to a patient.
Also some additional cleanup of the text in the "Add Vital Signs" form (bolding, spacing, clearing out metric units if no American units, etc).