Orders & Drugs
Added more values to the insert of Order Status about the order (dose,
calculated dose, etc).

Updated the 3 template database script to include the changes for the
Orders table.

Created a new VistA Order Entry script.

Created a new VistA script to pull the list of drugs from VistA's
National Library Medicine list and update the database Lookup Type with
the new list and the removal of old entries. The new entry description
is NLM. The Drug selection interface will need to be updated to remove
Outpatient and Inpatient and only select from one list, NLM.