Renamed the "Select Template currently applied to this patient" radio button and eliminated the drop down which would allow selection of current template. Since there can only be a single current template, clicking this button now displays the current template (eliminates multiple clicks to select from dropdown) "Select an existing standard template" works as before.
Display Template Information in CTOS Tab
At the bottom of the template displayed in the CTOS Tab a new listing of the # of Patients currently undergoing treatment with this regimen is listed.
The # of patients is a hot link which, when clicked, will display a popup window list of all the patients and their information:
Patient Name, Date Treatment Started, Projected End Date, SSID.

A new defect was identified:
When displying a template if there is no link for any clinical reference material a "Link to the PMID" anchor would still appear but that link opened up a new instance of COMS in a separate window.
This defect was fixed. Now, if there's no link information available the "Link to the PMID" does not appear.

Template List Tab - Centerred the # in the # of Patients column
Added Template Description (e.g. COMS Testing) to the popup window title (using same code module as in the display of a Template)
Added "Projected End Date" column to the listing.