Defect identified - Noticed within OEM that order status was properly displayed for Therapy medications, but not Pre- or Post-Therapy.  Therapy Meds also correctly (for today/dispensed meds) no longer enabled options of Cancel, Edit, or Hold.  Pre & Post both did…dispensed meds shouldn’t have those options & the previously Held Mylanta was no longer showing the Held status (I had manually refreshed for some reason during the past two steps).  There are two screenshots for this to show OEM and Orders Tab.
----------------
A typo in which block of medications to look through for the med to hold (and hence cancel as well) caused the code to look into the wrong section...
The test was done with the 4th medication of the Post Therapy section. But the typo caused the code to look for the 4th medication in the THERAPY section.
Since the Therapy section only had 3 medications the error was generated.